python/wxNLP/requirements.txt
certifi==2017.4.17
chardet==3.0.4
idna==2.5
olefile==0.44
Pillow==4.2.0
pypng==0.0.18

requesPyQRCode==1.2.1ts==2.18.1
urllib3==1.21.1
